chore: Move fe_from_str to integr tests
It doesn't make much sense to have a `utils` module which in reality
just has one function that only integration tests require.

So, chore: Move `fe_from_str` to integration tests file removing code
from the lib itself.
