+++ src/lib.rs
@@ -14,8 +14,6 @@ pub mod secp256k1;
 pub mod secp256r1;
 pub mod secq256k1;
 
-pub mod utils;
-
 #[macro_use]
 mod derive;
